+++ CHANGELOG.md
@@ -21,6 +21,7 @@ This file is used to list changes made in each version of cookbook-openstack-ima
 * Add support for multiple rabbit mq hosts
 * Make auth_version to be v2.0 in configuration file
 * Set the default cinder version to be v2
+* stores in glance-api.conf configurable with node attribute
 
 ## 9.2.0
 * python_packages database client attributes have been migrated to the -common cookbook